fixed various tag/layer issues, shells are now counted for points, baskets pull in shells
