+++ Assets/Prefabs/TestPrefabs/Shell1.prefab
@@ -15,7 +15,7 @@ GameObject:
   - component: {fileID: -4907559344524557666}
   - component: {fileID: 1251018937882810490}
   - component: {fileID: -8527566764055479666}
-  m_Layer: 19
+  m_Layer: 6
   m_Name: Shell1
   m_TagString: Untagged
   m_Icon: {fileID: 0}
