+++ Assets/Prefabs/TestPrefabs/Shell2.prefab
@@ -15,7 +15,7 @@ GameObject:
   - component: {fileID: -8712081266196796292}
   - component: {fileID: -1697970504273160632}
   - component: {fileID: 9049070729588719912}
-  m_Layer: 19
+  m_Layer: 6
   m_Name: Shell2
   m_TagString: Untagged
   m_Icon: {fileID: 0}
