+++ Assets/Prefabs/TestPrefabs/Shell3.prefab
@@ -15,7 +15,7 @@ GameObject:
   - component: {fileID: -4890035128826575560}
   - component: {fileID: -3579646112258433055}
   - component: {fileID: 8840667562207461876}
-  m_Layer: 19
+  m_Layer: 6
   m_Name: Shell3
   m_TagString: Untagged
   m_Icon: {fileID: 0}
