+++ Assets/Scenes/BeachTest.unity
@@ -956,7 +956,7 @@ MonoBehaviour:
   dragStrength: 1
   draggableLayers:
     serializedVersion: 2
-    m_Bits: 524288
+    m_Bits: 576
   tideRolling: 1
 --- !u!1001 &2071014153
 PrefabInstance:
