+++ Assets/Scripts/Events/EventScripts/UFO/UFOEvent.cs
@@ -11,10 +11,14 @@ public class UFOEvent : GameEvent
     {
         if (RoundManager.Instance.currRoundActive) {
             isRunning = true;
-            spawnLocations = FindFirstObjectByType<OffscreenSpawns>().spawns;
-            int randSpawn = Random.Range(0, spawnLocations.Length);
-            GameObject ufo = Instantiate(UFOPrefab, spawnLocations[randSpawn].position, Quaternion.identity);
-            ufo.GetComponent<UFO>().myEvent = this;
+            OffscreenSpawns oss = FindFirstObjectByType<OffscreenSpawns>();
+            if (oss != null)
+            {
+                spawnLocations = FindFirstObjectByType<OffscreenSpawns>().spawns;
+                int randSpawn = Random.Range(0, spawnLocations.Length);
+                GameObject ufo = Instantiate(UFOPrefab, spawnLocations[randSpawn].position, Quaternion.identity);
+                ufo.GetComponent<UFO>().myEvent = this;
+            }
         }
         yield return null;
     }
