+++ Assets/Scripts/Level Scripts/Cherry/Cherry.cs
@@ -3,9 +3,6 @@ using System.Collections;
 
 public class Cherry : LevelPickup
 {
-    [HideInInspector] public bool ignoreBasketPull = false;
-    public int pointValue = 1;
-
     protected override void Awake()
     {
         base.Awake();
@@ -14,12 +11,5 @@ public class Cherry : LevelPickup
     protected override void Update()
     {
         base.Update();
-    }
-
-    public IEnumerator TemporarilyIgnoreBasket(float duration)
-    {
-        ignoreBasketPull = true;
-        yield return new WaitForSeconds(duration);
-        ignoreBasketPull = false;
     }
 }
