+++ ProjectSettings/TagManager.asset
@@ -41,13 +41,13 @@ TagManager:
   - NoBackground
   - 
   - 
-  - Cherry
+  - 
   - (EFFECT)LayerMask
   - (EFFECT)Seethrough
   - Ground
   - SpecialGround
   - Killable
-  - OceanDraggable
+  - 
   - 
   - 
   -
